requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
pandas>=2.0.0
Pillow>=10.0.0
pypdf>=4.0.0
